fixed shadows. Added instead of multiplied. Sigh...

+++ OpenGLCourseApp/Window.h
@@ -13,8 +13,8 @@ class Window
 
     int initialize();
 
-    GLfloat getBufferWidth() { return bufferWidth; }
-    GLfloat getBufferHeight() { return bufferHeight; }
+    GLint getBufferWidth() { return bufferWidth; }
+    GLint getBufferHeight() { return bufferHeight; }
 
     bool getShouldClose() { return glfwWindowShouldClose(mainWindow); }
